Add persistence package
Create a persistence package for database integration using Ent. Define
database schema with a single `sip` table with `name` and `checksum`
columns that will be used to check for duplicates at the beginning of
the preprocessing workflow. The persistence layer will only be used when
`checkDuplicates` is set to `true` in the configuration.

[skip-codecov]