Check for duplicated SIPs
In the preprocessing workflow, expect a compressed SIP instead
of a directory. If checkDuplicates is enabled in the configuration,
calculate the SIP checksum and persist the SIP name and checksum
in the DB. If there is already an entry with the same checksum, fail
the workflow with a validation error.

- Add activity to calculate the SIP checksum.
- Add local activity to check for duplicates.
- Include archiveextract activity from temporal-activities.
- Add events for the three activities to the result.