Add file validator scope
This was prompted by work on #152 but is mostly unrelated.

This change recognizes that veraPDFValidator operates on a directory,
rather than an individual file, while allowing for future validators
that operate on a single file.

The added `fvalidate.Scope()` method returns the intended scope of the
`Validate()` method, either a single file or a directory of files.

The added 'veraPDFValidator.scope' property indicates that the validator
operates on a directory of files.