Simplify veraPDF validator creation
This change was prompted by work on #152, but is mostly unrelated.

I've simplified the `NewVeraPDFValidator` constructor by removing the
`runFunc` and `logger` parameters.

The `runFunc` parameter was solely used to avoid the dependency on the
veraPDF Java executable in test, but didn't provide any assurance that
the code would run correctly in production.

The `logger` parameter was used to log the exit code and STDERR when
the veraPDF executable returned an unexpected error code. Rather than
logging the veraPDF error data directly, the new code returns a custom
`SystemError` that includes the validator name, exit code, STDERR
message, and a human readable message intended for display in the Enduro
UI.