Conditionally add file validator PREMIS agents
Refs #152.

Don't add a PREMIS agent for a file validator when it doesn't validate
any files in the SIP. A validator will not validate files that don't
match its known file formats.

Because we are adding the validator agent to the PREMIS file before
we write all the PREMIS events, and the ordering is important in PREMIS,
we have to do some extra work to make sure we insert new events in the
right order.